Determine branch type using prefixes
When the `git branch` command is run, some branches have a prefix that
signal what kind of branch they are. The branch prefixed with `*` is the
branch that is currently checked out, while the `+` prefix is a branch
that is checked out in a Git worktree.

I'm using these prefixes to determine the branch type and then removing
them from the name. The type can be used later on to change styling when
showing to the user, since it is important for the user to know the
various branch types.