Add branch description to TUI
Since I have to manually set the delegate's height to 3, this only
currently supports branch descriptions that are a single line. To
support multi-line descriptions in the future, I may need a custom
delegate. I do not see a way to set the height for each individual item,
so changing it for specific items doesn't seem to be an option.