Add a simple, styled printout of the current branches
This styling basically matches the Python script that I use in my
dotfiles. I plan on making this tool more useful in the future using
some sort of TUI or something, but this simple printout with a
description, last commit, etc provides a good amount of useful
information for now.