Remove commit information from branch output
Originally I thought this would be useful, but it really just seems to
add a ton of clutter that isn't really necessary. If commit info *is*
needed in the future maybe it can be implemented as an optional flag or
config, but showing the branch, type, and description by default seems
like it should be good for 99% of calls.