Allow individual commands to handle errors
The error messages before could end up looking a bit weird with the
errors being handled in the main function, so having the individual
commands handle them and then just return an exit code should help make
this a lot cleaner.

Commands can now handle errors specific to their implementation (e.g.
the delete command needs to handle when the user quits the prompt) and
the fact that they return an exit code makes it more obvious in the main
function that once the commands are run the program exits.