Preserve original branch name sorting
Git will output the branch names sorted and since we get the branch
names directly from the output of Git, our list is also sorted. Before,
we were looping over the map to generate the list of choices to pass to
the prompt, but the map is unordered which causes the sort to be lost.

This just generates the list of choices earlier using the order of the
branches that are passed to the delete command handler.