Use pagination to allow for longer branch lists
Before the TUI would try to always list all of the branches out at once.
If you had a lot of branches, they would be listed off of the screen and
you wouldn't be able to see the cursor.

This creates a new list model that uses pagination to allow you to
scroll through long lists of branches.